+++ Graveyard Keeper/src/INeedSticks/GlobalUsings.cs
@@ -4,6 +4,7 @@ global using System.Collections.Generic;
 global using System.Globalization;
 global using System.Reflection;
 global using BepInEx;
+global using BepInEx.Configuration;
 global using BepInEx.Logging;
 global using HarmonyLib;
 global using Shared;