+++ Graveyard Keeper/src/NoIntros/GlobalUsings.cs
@@ -1,7 +1,9 @@
-﻿// Global using directives
+// Global using directives
 
 global using System.Reflection;
 global using BepInEx;
+global using BepInEx.Configuration;
 global using BepInEx.Logging;
 global using HarmonyLib;
 global using LazyBearGames.Preloader;
+global using Shared;