+++ Graveyard Keeper/src/FogBeGone/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 3.4.12 | 17 April 2026
+
+- Added an update notice on the main menu that flags when this mod (or others in this collection) has a newer version on Nexus. Click an entry to open its Nexus page. Toggle off in settings if you'd rather not see it
+
 ## 3.4.11 | 14 April 2026
 
 - Fixed constant fog indoors and outdoors caused by inverted weather logic