+++ Graveyard Keeper/src/NoIntros/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 2.2.9 | 17 April 2026
+
+- Added a quiet update notification on the main menu showing which mods in this collection have newer versions on Nexus. Click an entry to open its Nexus page. Opt-out via the new "Check for Updates" setting
+
 ## 2.2.8 | 11 April 2026
 
 - Added startup logging to help diagnose mod and game environment issues